:boom: refactor for QuerySet and QueryExecutor
